Log persistent CSV of device requests and metadata (#11)
Gets logged to a `persist/` folder and includes data like battery
voltage and reset status. The CSV can be obtained through the web
interface with a password, if the configuration is set up for it.

Also changes the configuration infrastructure so that devices is just an
entry, instead of the root, and can coexist with other fields like
password. Old style config will still load.